requirements-fe.txt
gradio==5.49.1
pillow==11.1.0
requests==2.32.3